Add build system infrastructure.
* Set output folder via environment variable.
* Add tar generation.
* Make builder work both in release and debug profiles.
* Add version handling.
* Avoid deep cloning Strings by using Rc<str>.
