+++ builder/src/bin/dummy.rs
@@ -1,0 +1,4 @@
+// Copyright 2024-Present Datadog, Inc. https://www.datadoghq.com/
+// SPDX-License-Identifier: Apache-2.0
+
+pub fn main() {}
