+++ tools/docker/Dockerfile.build
@@ -112,12 +112,14 @@ COPY "data-pipeline/Cargo.toml" "data-pipeline/"
 COPY "data-pipeline-ffi/Cargo.toml" "data-pipeline-ffi/"
 COPY "bin_tests/Cargo.toml"  "bin_tests/"
 COPY "tinybytes/Cargo.toml" "tinybytes/"
+COPY "builder/Cargo.toml"  "builder/"
 RUN find -name "Cargo.toml" | sed -e s#Cargo.toml#src/lib.rs#g | xargs -n 1 sh -c 'mkdir -p $(dirname $1); touch $1; echo $1' create_stubs
 RUN echo \
     bin_tests/src/bin/crashtracker_bin_test.rs \
     bin_tests/src/bin/crashtracker_receiver.rs \
     bin_tests/src/bin/crashtracker_unix_socket_receiver.rs \
     bin_tests/src/bin/test_the_tests.rs \
+    builder/src/bin/dummy.rs \
     ddtelemetry/examples/tm-worker-test.rs \
     ipc/benches/ipc.rs \
     ipc/tarpc/tarpc/examples/compression.rs \
